chore(tests): moved `GameStorageTest` into an own file
